[react-events] Split out mixed event responder tests
Fix
