+++ packages/react-events/src/dom/__tests__/Press-test.internal.js
@@ -1120,334 +1120,7 @@ describe.each(environmentTable)('Press responder', hasPointerEvents => {
     target.pointerdown();
   });
 
-  it('should correctly pass through event properties', () => {
-    const timeStamps = [];
-    const ref = React.createRef();
-    const eventLog = [];
-    const logEvent = event => {
-      const propertiesWeCareAbout = {
-        pageX: event.pageX,
-        pageY: event.pageY,
-        screenX: event.screenX,
-        screenY: event.screenY,
-        clientX: event.clientX,
-        clientY: event.clientY,
-        pointerType: event.pointerType,
-        target: event.target,
-        timeStamp: event.timeStamp,
-        type: event.type,
-      };
-      timeStamps.push(event.timeStamp);
-      eventLog.push(propertiesWeCareAbout);
-    };
-
-    const Component = () => {
-      const listener = usePress({
-        onPressStart: logEvent,
-        onPressEnd: logEvent,
-        onPressMove: logEvent,
-        onPress: logEvent,
-      });
-      return <button ref={ref} listeners={listener} />;
-    };
-    ReactDOM.render(<Component />, container);
-
-    const target = createEventTarget(ref.current);
-    target.setBoundingClientRect({x: 10, y: 10, width: 100, height: 100});
-    target.pointerdown({
-      pointerType: 'mouse',
-      pageX: 15,
-      pageY: 16,
-      x: 30,
-      y: 31,
-    });
-    target.pointermove({
-      pointerType: 'mouse',
-      pageX: 16,
-      pageY: 17,
-      x: 31,
-      y: 32,
-    });
-    target.pointerup({
-      pointerType: 'mouse',
-      pageX: 17,
-      pageY: 18,
-      x: 32,
-      y: 33,
-    });
-    target.pointerdown({
-      pointerType: 'mouse',
-      pageX: 18,
-      pageY: 19,
-      x: 33,
-      y: 34,
-    });
-    expect(typeof timeStamps[0] === 'number').toBe(true);
-    expect(eventLog).toEqual([
-      {
-        pointerType: 'mouse',
-        pageX: 15,
-        pageY: 16,
-        screenX: 30,
-        screenY: 81,
-        clientX: 30,
-        clientY: 31,
-        target: ref.current,
-        timeStamp: timeStamps[0],
-        type: 'pressstart',
-      },
-      {
-        pointerType: 'mouse',
-        pageX: 16,
-        pageY: 17,
-        screenX: 31,
-        screenY: 82,
-        clientX: 31,
-        clientY: 32,
-        target: ref.current,
-        timeStamp: timeStamps[1],
-        type: 'pressmove',
-      },
-      {
-        pointerType: 'mouse',
-        pageX: 17,
-        pageY: 18,
-        screenX: 32,
-        screenY: 83,
-        clientX: 32,
-        clientY: 33,
-        target: ref.current,
-        timeStamp: timeStamps[2],
-        type: 'pressend',
-      },
-      {
-        pointerType: 'mouse',
-        pageX: 17,
-        pageY: 18,
-        screenX: 32,
-        screenY: 83,
-        clientX: 32,
-        clientY: 33,
-        target: ref.current,
-        timeStamp: timeStamps[3],
-        type: 'press',
-      },
-      {
-        pointerType: 'mouse',
-        pageX: 18,
-        pageY: 19,
-        screenX: 33,
-        screenY: 84,
-        clientX: 33,
-        clientY: 34,
-        target: ref.current,
-        timeStamp: timeStamps[4],
-        type: 'pressstart',
-      },
-    ]);
-  });
-
   if (hasPointerEvents) {
-    // TODO(T46067442): move these tests to core
-    /*
-    it('should properly only flush sync once when the event systems are mixed', () => {
-      const ref = React.createRef();
-      let renderCounts = 0;
-
-      function MyComponent() {
-        const [, updateCounter] = React.useState(0);
-        renderCounts++;
-
-        function handlePress() {
-          updateCounter(count => count + 1);
-        }
-
-        const listener = usePress({
-          onPress: handlePress,
-        });
-
-        return (
-          <div>
-            <button
-              ref={ref}
-              listeners={listener}
-              onClick={() => {
-                updateCounter(count => count + 1);
-              }}>
-              Press me
-            </button>
-          </div>
-        );
-      }
-
-      const newContainer = document.createElement('div');
-      const root = ReactDOM.unstable_createRoot(newContainer);
-      document.body.appendChild(newContainer);
-      root.render(<MyComponent />);
-      Scheduler.unstable_flushAll();
-
-      const target = ref.current;
-      target.dispatchEvent(pointerdown({timeStamp: 100}));
-      target.dispatchEvent(pointerup({timeStamp: 100}));
-      target.dispatchEvent(click({timeStamp: 100}));
-
-      if (__DEV__) {
-        expect(renderCounts).toBe(2);
-      } else {
-        expect(renderCounts).toBe(1);
-      }
-      Scheduler.unstable_flushAll();
-      if (__DEV__) {
-        expect(renderCounts).toBe(4);
-      } else {
-        expect(renderCounts).toBe(2);
-      }
-
-      target.dispatchEvent(pointerdown({timeStamp: 100}));
-      target.dispatchEvent(pointerup({timeStamp: 100}));
-      // Ensure the timeStamp logic works
-      target.dispatchEvent(click({timeStamp: 101}));
-
-      if (__DEV__) {
-        expect(renderCounts).toBe(6);
-      } else {
-        expect(renderCounts).toBe(3);
-      }
-
-      Scheduler.unstable_flushAll();
-      document.body.removeChild(newContainer);
-    });
-
-    it('should properly flush sync when the event systems are mixed with unstable_flushDiscreteUpdates', () => {
-      const ref = React.createRef();
-      let renderCounts = 0;
-
-      function MyComponent() {
-        const [, updateCounter] = React.useState(0);
-        renderCounts++;
-
-        function handlePress() {
-          updateCounter(count => count + 1);
-        }
-
-        const listener = usePress({
-          onPress: handlePress,
-        });
-
-        return (
-          <div>
-            <button
-              ref={ref}
-              listeners={listener}
-              onClick={() => {
-                // This should flush synchronously
-                ReactDOM.unstable_flushDiscreteUpdates();
-                updateCounter(count => count + 1);
-              }}>
-              Press me
-            </button>
-          </div>
-        );
-      }
-
-      const newContainer = document.createElement('div');
-      const root = ReactDOM.unstable_createRoot(newContainer);
-      document.body.appendChild(newContainer);
-      root.render(<MyComponent />);
-      Scheduler.unstable_flushAll();
-
-      const target = ref.current;
-      target.dispatchEvent(pointerdown({timeStamp: 100}));
-      target.dispatchEvent(pointerup({timeStamp: 100}));
-      target.dispatchEvent(click({timeStamp: 100}));
-
-      if (__DEV__) {
-        expect(renderCounts).toBe(4);
-      } else {
-        expect(renderCounts).toBe(2);
-      }
-      Scheduler.unstable_flushAll();
-      if (__DEV__) {
-        expect(renderCounts).toBe(6);
-      } else {
-        expect(renderCounts).toBe(3);
-      }
-
-      target.dispatchEvent(pointerdown({timeStamp: 100}));
-      target.dispatchEvent(pointerup({timeStamp: 100}));
-      // Ensure the timeStamp logic works
-      target.dispatchEvent(click({timeStamp: 101}));
-
-      if (__DEV__) {
-        expect(renderCounts).toBe(8);
-      } else {
-        expect(renderCounts).toBe(4);
-      }
-
-      Scheduler.unstable_flushAll();
-      document.body.removeChild(newContainer);
-    });
-
-    it(
-      'should only flush before outermost discrete event handler when mixing ' +
-        'event systems',
-      async () => {
-        const {useState} = React;
-
-        const button = React.createRef();
-
-        const ops = [];
-
-        function MyComponent() {
-          const [pressesCount, updatePressesCount] = useState(0);
-          const [clicksCount, updateClicksCount] = useState(0);
-
-          function handlePress() {
-            // This dispatches a synchronous, discrete event in the legacy event
-            // system. However, because it's nested inside the new event system,
-            // its updates should not flush until the end of the outer handler.
-            button.current.click();
-            // Text context should not have changed
-            ops.push(newContainer.textContent);
-            updatePressesCount(pressesCount + 1);
-          }
-
-          const listener = usePress({
-            onPress: handlePress,
-          });
-
-          return (
-            <div>
-              <button
-                listeners={listener}
-                ref={button}
-                onClick={() => updateClicksCount(clicksCount + 1)}>
-                Presses: {pressesCount}, Clicks: {clicksCount}
-              </button>
-            </div>
-          );
-        }
-
-        const newContainer = document.createElement('div');
-        document.body.appendChild(newContainer);
-        const root = ReactDOM.unstable_createRoot(newContainer);
-
-        root.render(<MyComponent />);
-        Scheduler.unstable_flushAll();
-        expect(newContainer.textContent).toEqual('Presses: 0, Clicks: 0');
-
-        const target = button.current;
-        target.dispatchEvent(pointerdown({timeStamp: 100}));
-        target.dispatchEvent(pointerup({timeStamp: 100}));
-        target.dispatchEvent(click({timeStamp: 100}));
-
-        Scheduler.unstable_flushAll();
-        expect(newContainer.textContent).toEqual('Presses: 1, Clicks: 1');
-
-        expect(ops).toEqual(['Presses: 0, Clicks: 0']);
-      },
-    );
-    */
     it('should work correctly with stopPropagation set to true', () => {
       const ref = React.createRef();
       const pointerDownEvent = jest.fn();
